Split endpoints and db code into individual files
